split LRParserGenerator in LRFirstFollow, LRAutomaton, LR0Closure and LRParserGenerator
