+++ src/main/scala/pudu/parser/generator/LR1ParserGenerator.scala
@@ -3,15 +3,16 @@ package pudu.parser.generator
 import pudu.grammar._
 import pudu.parser._
 
-class LR1ParserGenerator[Tree, Token <: scala.reflect.Enum](grammar: Grammar[Tree,Token]) extends LRParserGenerator(grammar):
+class LR1ParserGenerator[Tree, Token <: scala.reflect.Enum](grammar: Grammar[Tree,Token]) extends ParserGenerator[Tree, Token]:
+  type StateT = State[Tree, Token]
 
-  import grammar._
+  val augmentedGrammar = grammar.augmented
+  import augmentedGrammar.{rules, startRules, eof, isNonTerminal}
 
-  /** start state */
-  override val startState = closure(startRules.map(_.toItem(eof)))
+  val lrff = LRFirstFollow(augmentedGrammar)
 
   /** state closure for LR1 parsers. */
-  override def closure(state: StateT): StateT =
+  def closure(first: Map[Symbol, Set[Terminal[Token]]])(state: StateT): StateT =
     def firstSeq(seq: Seq[Symbol], next: Option[Symbol]) =
       if !seq.isEmpty then first(seq.head)
       else if next.isDefined then first(next.get)
@@ -32,11 +33,16 @@ class LR1ParserGenerator[Tree, Token <: scala.reflect.Enum](grammar: Grammar[Tre
       else closureImpl(acc ++ step)
     closureImpl(state)
 
-  lazy val reduceActions: Map[(StateT, Terminal[Token]), Set[RuleT]] =
+  /** start state */
+  val startState = closure(lrff.first)(startRules.map(_.toItem(eof)))
+
+  val lra = LRAutomaton(startState, closure(lrff.first))
+
+  val reduceActions: Map[(StateT, Terminal[Token]), Set[Rule[Tree, Token]]] =
     /* First, we generate tuples (state, terminal, rule), such that there exists
      * an item X -> \alpha\cdot in state, and terminal\in FOLLOW(X). */
     val reduceByCases = for
-      state <- indexedStates.keys
+      state <- lra.indexedStates.keys
       item <- state
       if item.after.isEmpty && item.context.isDefined
     yield (state, item.context.get, item.rule)
@@ -45,3 +51,4 @@ class LR1ParserGenerator[Tree, Token <: scala.reflect.Enum](grammar: Grammar[Tre
     reduceByCases.groupMap(t => (t._1, t._2))(_._3).view
       .mapValues(_.toSet).toMap
 
+  def parser = LRParserGenerator(augmentedGrammar, lra.lrAutomaton, lra.indexedStates, reduceActions).parser
