+++ src/test/scala/pudu/parser/generator/LRReportTest.scala
@@ -6,7 +6,7 @@ import scala.util.matching._
 // import org.scalatest.mockito.MockitoSugar // no Scala 3 support :(
 // import org.mockito.Mockito._
 
-class LRReportTest extends munit.FunSuite {
+class LRReportTest extends munit.FunSuite {/*
   val gen = SLRParserGenerator(SimpleArithmetic.grammar.augmented)
   val report = LRReport(gen)
 
@@ -111,6 +111,6 @@ class LRReportTest extends munit.FunSuite {
                            template.replace("<v1>", "4").replace("<v2>", "3"))
     assert(alternatives.contains(report.all))
   }
-
+*/
 
 }
